Changes to allow path pricing in RestOre
Biggest change is in scenariosimmarket where a constructor is added that
uses a base scenario instead of an initial market. This is not finished
but added to see if we want to go this way.

Another price function has been added to the Static class in pricer.hpp.
This should price a portfolio on scenario paths and return a cube.