Add flag for turning on and off Equity curve simulation
Need to be able to turn on and off equity forecast and dividend curve
simulation. The current use case is in general:
- "simulate" equity forecast and dividend curves when doing sensitivity.
  In other words, allow for shifting these curves.
- do not simulate equity forecast and dividend curves when computing
  exposure.

Added the SimulateEquityNames node to Equities node in simulation XML
file for turning this on and off as for the other risk factors. By
default, it is set to false.