Without this, docker build of ORE+ fails
Docker first builds ORE and _installs_ it. It then deletes the source
tree. lib files go to /usr/local/lib and headers go to
/usr/local/include as usual. Without this change, the version.hpp files
are not added to /usr/local/include. This leads to the error below when
attempting to build ORE+ on top of the _installed_ ORE.

In file included from /usr/local/include/ored/ored.hpp:8:0,
                 from /usr/local/include/orea/app/oreapp.hpp:34,
                 from ../../oreap/app/oreappplus.hpp:12,
                 from oreappplus.cpp:10:
/usr/local/include/ored/version.hpp:25:27: fatal error: qle/version.hpp: No such file or directory
 #include <qle/version.hpp>
modified:   OREAnalytics/orea/Makefile.am