Fix examples broken by Example 26
Not sure any more of the usefulness of the '_flat.txt' market file as
non flat data has been copied in to it over time. Maybe we should just
have a single market data file for examples _OR_ use separate
configurations if we wish to keep a flat market also.