Shift sizes in the sensitivity reports are absolute shifts in the given risk factor key
Here, every time a scenario description is added for an 'up' shift, a
default shift of 0.0 (looks better than Null<Real>() in reports). Later,
if a valid shift size can be caculated, the value is overwritten with a
valid shift size for the risk factor key.

A valid shift size is harder to calculate if the shift is as a result of
an interpolated value i.e. if sim market setup does not line up with the
sensitivity market set up. Just left it as a default value of 0.0 here.

Might be better just to return two columns in the reports i.e. shift
size and shift type instead of the absolute shift in the risk factor.
Don't think that the absolute shift calculated and output to the report
is used anywhere.