Add a filtered scenario streaming class
This allows us to only stream scenarios where the delta or gamma or
non-negligible and to ignore the rest.

Important in RESTORE because there we stream from a cube and without
this we get all the sensitivity records, even those where delta and
gamma are 0. This is not the case in ORE+ because we run the SIMM
backtest after loading the sensitivities from file and the records are
filtered at the file writing stage with a non-zero threshold generally.