Error when parsing XML file on Windows using VS 15.8.1
This is a fix to correct an error encountered when parsing XML files on
Windows using Visual Studio 15.8.1.

When the length of the buffer is calculated on Windows seekg and tellg,
the eol character \r\n is counted as 2. Then when the file is read in to
the buffer using read, the number of characters read, as reported by
calling gcount on the ifstream, is counting the eol character as 1.
Therefore, the buffer length is longer than the number of characters
read. This was leading to garbage between the gcount element of the
buffer and the end of the buffer i.e. the length element of the buffer.
This then lead to parsing errors when a call was made to parse in
RapidXml. Setting the null terminating character at gcount avoids this.