Utility macros to ease use of files in tests
These macros will work as long as we have a flat file layout in the
'test' folder for the various libraries. The macros expect that input
files for a test in source files 'source.xpp' are added in the directory
'test/input/source' and any output will be sent to 'test/output/source'.

If we start using a folder structure for the source files under test, we
would have to make the macros more elaborate.