Providing a base data path to the unit tests
Could not use what I had before as I was falsely assuming that __FILE__
would expand to the **full** directory path of the file during
preprocessing.

Instead, here, you need to provide a base path parameter in order to run
the test suite. In other words, the minimal command line run is now:
  OREData-test-suite-vc141-x64-mt.exe -- --base_data_path=..\
on Windows for example.