[touch: 711]Amendments for COP OIS swap and regression tests
The assumptions underlying the regression tests are in the Readme.md
files. The market COP OIS swap quotes may be non-deliverable and settled
in USD and may also be collateralised in USD. This would require a
rethink of how we use them. In these regression tests, we assume that
the market quotes are collateralised and settled in COP.