Allow user to override tenor on one leg of cross currency helper
The use case for this is MXN 28D vs USD 1M cross currency basis swaps.
For these swaps, the roll on the USD leg is every 28D with USD-LIBOR-1M
being paid every 28D. This change allows the swap underlying the helper
to be built with the correct schedule on the USD leg.