Update ConfigurationBuilder
There is some tidy up here and some changes:
- all sensitivity configurations in one place i.e. in method shiftPair.
  We have updated the configuration here to have absolute shift of 1%
  for FX, commodity and equity volatility if the run type is SIMM.
- all tenors are taken from one place i.e. the method simmTenors and are
  explicitly specified in static members for clarity
- par conversion is only needed if the run type is SIMM

Still more additions to make i.e. make sure all enums are covered where
they are used with an error message if an implementation is not yet
there for them.

Check that the inflation and cap floor tenors do not give errors in the
par sensitivity analysis after pulling in a fix from a branch from
Peter. Before now, the short dated tenors would give an error because
the par instruments would have no sensitivity to them.