Add a piecewise optionlet stripper
This stripper takes input cap floor volatilities or premiums via the
CapFloorHelper instruments and bootstraps the underlying optionlet curve
using a given interpolation method.

The stripping happens along a strike column. Next step will be a
stripper that holds a vector of these objects in order to strip a
surface and derives from StrippedOptionletBase.