[touch: 721]What happens between time 0 and first optionlet time
Iterative bootstrap requires the firstDate on the curve to be the
reference date i.e. time 0. For optionlet curves, it may make more sense
to have the first date as the first optionlet fixing date where we have
implied an optionlet volatility and leave the rest up to some
extrapolation method. We are forced then to explicitly state what
happens between time 0 and the first time for which we imply an
optionlet volatility. The flatFirstPeriod flag indicates that the
optionlet volatility is flat from time 0 out to first optionlet fixing
date where we have implied a volatility. If it is false, the time 0
optionlet volatility is set to 0 and the curve's interpolation method
gives the intermediate volatilities.