[touch: 721]Allow tests of cubic interpolation in time direction pass
Added a QuantExt::IterativeBootstrap that is a straight copy of the
QuantLib version except that there is a global accuracy parameter that
is less strict than the accuracy in the Curve being bootstrapped.

May need to add more items to the QuantExt::IterativeBootstrap e.g. if
the solver fails in a section, set the data in that section to a minimum
value or a previous value and move to the next section. Will look at
this later.

Open question how to pass non-default constructed Bootstrap in
PiecewiseOptionletStripper down to each of the strikeCurves_. Running in
to issues here.