[touch: 741]Only allow multiple index names in the portfolio XML
We should be using the single unique name for an ibor index in the
configuration, market data and fixing files. In other words, both
elements in the following list will work if input in a portfolio XML but
only the first will work in our configuration, market data and fixing
files:
- USD-FedFunds, USD-FedFunds-1D
- EUR-EONIA, EUR-EONIA-1D
- CAD-CDOR-3M, CAD-BA-3M
- EUR-EURIBOR-3M, EUR-EURIB-3M
- USD-SIFMA, USD-SIFMA-1W, USD-SIFMA-7D
etc.
The alternative names in the portfolio XML get converted to the single
unique name during the portfolio loading.