[touch: 721]Allow for non-relative date based cap floor helpers
Allow CapFloorHelper not respond to changes in the evaluation date.
Still derives from RelativeDateBootstrapHelper as too much duplication
and to make another CapFloorHelper that derives from BootstrapHelper and
call it something else.

Updated the PiecewiseAtmOptionletCurve class and tests for moving and
non-moving term structure.

The PiecewiseOptionletStripper class needs to continue to have relative
date based helpers as it derives from OptionletStripper which only
allows a relative date based CapFloorTermVolSurface in its construction.
To see this, it registers with the global evaluation date and calls
termVolSurface_->settlementDays() which will not be implemented for a
non moving termstructure. This settlementDays method is then call when
the stripper is passed to any of the adapters. It is all set up for only
relative date based strippers and adapters. A lot of work to cover fixed
date based variants.