[touch: 721]This is the old setup that makes the unit tests pass
This configuration does not make much sense as a default i.e. cubic
interpolation on the term surface and then linear on the optionlet
surface. More sensible to have the same on both as a default. This
commit will be reverted - only recording it here to show that the unit
tests pass.