[Touch: 783]Update the ORE piece to have a leg data factory
The doxygen comments in legdatafactory.hpp explain its use. The idea is
based on the post at: https://stackoverflow.com/a/582456/1771882. We
have a leg data factory and each class deriving from LegAdditionalData
must register with this factory. As long as a leg data type registers
with the factory, one can call build on the factory and it will return
an instance of that leg data type.

As mentioned in the ticket, LegData::initialiseConcreteLegData is now
much cleaner.

Also, we should never have touch LegData or LegDataFactory when adding
new classes to the LegAdditionalData hierarchy. We just have to register
them as described in the comments for the LegDataRegister class. This
merely involves adding a static member to the new class deriving from
LegAdditionalData.

Next step remove LegDataPlus in ORE+ and integrate it with this.