[Touch: 783]Refactor to avoid seg fault on Linux
Believe that having LegDataFactory as a static member in LegData was
triggering a seg fault on Linux. It could be the same seg fault that we have
seen before on Linux coming from Boost Serialisation. There is an email
thread on this Fri 21 Jun with subject "Boost issue with 1.65.1 to 1.68
on Linux". There are issues open with Boost also at:
https://github.com/boostorg/serialization/issues/104
https://svn.boost.org/trac10/ticket/13186
When I ran in gdb and did a backtrace, I was pointed to
boost::serialization::singleton.

This refactor avoids the issue. If there is an issue with LegDataFactory
being a Singleton, I can go beack to the suggestion in the ticket and
have a static map in LegData that we add builders to with a static
method but I don't see much gain.