[Touch: 843]QuantExt::IterativeBootstrap update interp on last
This fixes an issue that was coming up with optionlet bootstraping in
AS. It showed up when dontThrow was set to true. If the bootstrap is
still failing on the final alive instrument, the underlying curve's
interpolation object does not have update() called. You then see the
last used value in the attempted solve showing up when you ask the curve
for a value around that point.