[Touch: 843]Use a better fallback when dontThrow is true
If dontThrow is true in the IterativeBootstrap, we find a better
fallback value to use in the event that the bootstrap fails on a given
pillar. Before, we just used the previous pillar's value. The change
here searches between the min and the max, that was used in the current
pillar's root find, to find the value that yields the minimum bootstrap
error in absolute value. In most cases, this will be either the min or
max value itself.