[Touch:996]Future expiry calc gives option expiries + new anchor rule
We will need the ability to derive option expiry dates when we are
loading commodity option surfaces later so add it here and test it.

Added a new anchor day rule to the commodity future conventions to deal
with the old ICE Brent expiry rules. We will need this when loading
commodity volatility surfaces during backtesting.