[touch:1073]fix shift multiplier for cap/floor vols
This is not only a fix, but seems to greatly simplify the code,
something we could do for the other risk factors as well. Do this
another time and separately from this fix here though not to avoid a
too big change at once in things that are unrelated to the original
problem.