[Touch:904]Update CDS vol curve and market data to allow for strikes
Allow for CDS option surfaces as well as curves and constant volatility
structures. Allow for wild cards in the specification as for equity and
commodity. Make the constant and curve variant more robust in that there
are explicit quotes and errors where duplicates.

Left Expiries node in cds vol configuration to be backwards compatible.
Shouldn't be used because it is likely to cause errors where the market
data has multiple quotes for a given expiry.

Have a StrikeType in the cds vol configuration. Not using it yet because
its use depends on other outstanding tickets i.e.
https://qrm.codebasehq.com/projects/openxva/tickets/691
Can align after this ticket is completed. Main thing to guard against is
asking a strike _price_ based cds vol structure for a volatility at a
given strike _spread_ and vice versa. Will assume for now that the
market data and ORE trade XML is consistent in this respect.