[Touch:904]Differentiate option surfaces for different index terms.
Not happy with this but don't want to start implementing a cube for a
release that is due in 4 days when it has not been considered until this
point.

In future, we will need to decide if we are happy to continue with
separate option surfaces for different CDS index terms or if we want one
CDS index option cube which holds information on all terms. The market
is different from the IR swaption market in that we should only have CDS
index options on the predefined CDS index terms e.g. 3Y, 5Y, 7Y and 10Y.