Test that this fixes the build on dev114 before changing the rest.
Seems odd to be adding the dependencies at the the QuantExt level Cmake
file and not at the QuantExt test level Cmake file e.g. unit_test_framework
at the QuantExt level. Just copying what is there and assuming there was a
reason for it when the cmake files were created.

For the timer and chrono piece, I am using what is in QuantLib at:
https://github.com/lballabio/QuantLib/blob/master/test-suite/CMakeLists.txt

The movement of the "REQUIRED" flag in find_package is based on the
QuantLib commit and the cmake docs here:
https://cmake.org/cmake/help/latest/command/find_package.html