[touch:1170]add option to specify initialPrice in target ccy
Cassini tells us that this is the usual way to specify the initial
price for composite equity swaps in the term sheet, i.e. both the
initial equity price _and_ the fx rate that converts it to the trade
ccy are defined in the term sheet (instead of using an FX fixing).