[Touch:933]Fix to make xccy mtm regression test work both ways.
With this change the xccy mtm regression test gives the same results
with and without the xbs_par_conversion_off flag. This is a starting
point for the changes. Note that the crif.csv is identical to the
expected output. The simm scenario, sensitivity and par sensitivity
files have their results re-ordered when compared with the expected
output but the contents, i.e. the sensitivities, are identical.