Allow overriding of default VS code analysis rule set.
Some of the warnings are known and will not be changed and are
distracting. For example, the currently disabled warning about the use
of unscoped enums. We will not be correcting this and are content using
them.

May need to add a condition like the following in ore.props:
Condition="'$(VisualStudioVersion)' >= '16.0'"
Wait and see do others have issues on VS versions before 2019.