[Touch:1199]Document default curve bootstrapping from CDS spreads or prices.
Tidy up and add documentation for bootstrapping of CDS curves.

Also, add documentation for the new flag ImplyDefaultFromMarket which
when set to false gives the old behaviour but when set to true gives the
behaviour outlined in ticket 1199 as a workaround or temporary fix for
reference entities in the period between credit event determination date
and auction date.