[Touch:1244]Use latest version of latexmk in Docker pdf build
These changes allow us to use the latest version of latexmk in the Docker
pdf file build. In particular, there is the -Werror flag that treats
warnings as errors. Will have to put this on hold for now as there
are too many warnings to start dealing with. The hope was that this would
allow us to catch undefined references and citations but there are
too many others so the build would fail quite often.