Amend boost, ql and ore+ debug docker builds.
Leave the source files in /lib_src for gdb and valgrind. Deleted the
build directories created for cmake build at the end of the Docker
build to make the images smaller. The .so files have an
absolute path to the original source files in any case i.e. when
you look at `objdump -g shared_lib_name.so` and the DW_AT_name value
under compilation unit.