[Touch:1264]Update some example results and "fix" unit tests.
The unit tests will need to be revisited. The tolerances and steps were
chosen to just pass one test. When the test is generalised to run over
more inputs, the tests fail. Either the tolerances should be relaxed or
the steps increased (will probably be too slow). An example is the
existing DK martingale test. The Euler portion of this test fails if the
number of steps are _increased_ which suggests that for the smaller
number of steps, the existing pass is merely luck.