[touch:1372]skip indices of types other than EQ, COMM
The parseIndex() call will produce an exception when a convention
based ibor index is parsed - we don't need that at least for now,
since we only cover EQ, COMM indices here anyhow. It would be cleaner
to pass conventions to the parser, but we don't have access to them
here. I'd rather wait until we refactor ORE to have a convention
singleton (I think we want to do that), then the problem goes away
without any additional effort.