[Touch:1264]Add the extra piece to InfJyBuilder::requiresRecalibration()
The equivalent of volSurfaceChanged for the other builders. For InfDkBuilder,
the instrument underlying the calibration instrument is rebuilt from the
definition in the data, a pricing engine is attached and the NPV is
queried. I reuse the calibration helper here instead of rebuilding the
instrument - may be missing something and have to revisit but will try
it like this for now.