[Touch:1290]Quick partial fix for CDO with defaulted entities.
A lot more needs to be done on this when there is time. We at least here
adjust for defaulted entities up to the valuation date. We do not take
into account the timing of defaults in the past which is needed for
correct accruals in the current period and past cashflows. Also, may not
be the most elegant solution to have the default information in the
basket data or reference data but to have it elsewhere and create a link
would take a lot more design, coding and time.

Still need to add corrections to the base correlation structure building
in the market to:
- take account of changed attachment and detachment points
- take account of the tenor being from a fixed start date in the past

There are two examples:
- ExamplePlus_CDO_Reference_Data
- ExamplePlus_CDO_Basket_Data

They both show the valuation of all tranches of the CDX IG NA Series 33
(no defaults, index factor = 1.0) and all tranches of the CDX HY NA
Series 29 Version 14 (13 defaults, index factor = 0.87). The first
example uses reference data and the second example does not use
reference data and instead uses basket data in the trade directly.

The valuation of the CDX HY NA Series 29 Version 14 is incorrect because
it is reading the base correlations of the wrong points on the market
base correlation curve. As mentioned above, this will be fixed in a
subsequent commit.