[Touch:1269]Allow user to configure 2 sided deltas in sensi run.
This will be used with a moderately smaller shift size, e.g. 0.1% as
opposed to 1%, for risk factor key type FXSpot to get risks closer to
those expected by the theory for a CRIF run. This will make an
improvement generally before we even consider using the analytic FX
delta for those instruments where it is available.

The 2 sided delta is configurable in sensitivity data on the risk factor
key type.