Make calls for engines as late as possible
The engine calls add the market data elements during portfolio analysis.
Try to make them as late as possible so that if there are any simple
parsing errors, etc that cause trade building to fail, the market data
elements are not registered and there are potentially a lot few errors
downstream.