[Touch:1431]Inflation amendments to cover AU CPI like inflation
Quite specific to AUCPI as it stands. Allows for the addition of a
publication schedule and a publication roll option to the inflation
swap conventions. This is used to determine the start and end date of
the inflation swap helpers in line with the AFMA Inflation Product
conventions, May 2017. The unit tests check the behaviour of the curve
and the pricing of the market instruments around the Q3 release date of
28 Oct 2020.

Note that QL expects the fixings to be provided for the first day of the
period to which they apply. So, for example, the Q2 inflation index
fixing, i.e. the index value covering the period from 1 Apr 2020 to 30
Jun 2020 is expected with a date of 1 Apr 2020. In Refinitiv for
example, this index value is given with a date of 30 Jun 2020. Further
amendments needed on this ticket to cover that.