[Touch:1549]Take the fixed quantity from the floating leg if necessary.
Allow for omission of quantities on the commodity fixed leg. If they are
omitted, the quantities are taken from the commodity floating leg. This
avoids duplication of data and logic in the fixed leg i.e. PerHour,
PerPricingDay, etc. and counting the number of pricing dates in the
period. Also allow for addition of tag on the fixed and floating legs to
indicate the floating leg that the fixed leg should use. If tag is
empty, any floating leg is chosen.