[Touch:1641]Refactor EquityOptionSurfaceStripper to allow for commodity
Refactor EquityOptionSurfaceStripper to derive from abstract base class
OptionSurfaceStripper. Will add a CommodityOptionSurfaceStripper in
subsequent commits. The only methods that these derived classes need to
provide are 1) BS process and 2) forward price at given date.

Added a preferOutOfTheMoney flag to the stripper to indicate which
options to choose if both call and put available. For equity, set it to
false for backward compatibility with what is there before. Allow this
to be set via the equity vol curve configuration.

Added a new class Solver1DOptions in QLE and a corresponding XML
serialisable class OneDimSolverConfig in OREData so that options for the
sovler used in going from price to volatility can be passed in. For
backwards compatibility, added a default instance of this class to
equity vol curve config in case one is not provided via XML. The main
enhancement here is to be able to choose at XML level between a max-min
based search or a step based search. The solve method in the stripper is
bound based on these values.

In the stripper, if the price to vol conversion fails, the volatility is
not added to the sparse surface. This is a change in behaviour. Before
this, a vol of 0.0 was added.