[Touch:1641]Check to be reverted.
This most likely corrects the failing equity regression tests. Before,
when imply vol failed, the vol was set to 0 and used. After, when imply
vol fails, it is not used in the final surface which I feel is a better
approach. Will revert this change once I check that all regression
passes and update the failing regression tests.

The tests that are failing are nonsense. A "single name" equity option
on STOXX50E with strike ~12 when the spot for STOXX50E is at ~3750 in
the market data file. Not going to bother amending these.