[Touch:1677]Add helper for off-peak electricity futures.
The helper accounts for the fact that the monthly averaging off-peak
future and the daily off-peak future refer to different periods i.e. the
daily off-peak future is for 8 hours every day, the monthly averaging
future averages LMPs for the same 8 hours on good (NERC) business days
but averages LMPs over all 24 hours on (NERC) holidays and weekends.

Attached a workbook to the ticket showing how the price curves built for
the daily future contract projection reprice all input instruments.