[Touch:1687]Add commodityIndex method to Market.
Add this method so that pricing engines can request a commodity index
from the Market. This way the commodity index is created in the market
by a call to parseCommodityIndex where it has access to a Conventions
instance. The commodity index instance returned from the market can
later be cloned to change the expiry date (or the price curve). There
should be limited direct calls to the ctors of CommoditySpotIndex,
CommodityFuturesIndex and OffPeakPowerIndex throughout the code after
this. In general, creation of these should go through
parseCommodityIndex.