[Touch:1549]Use AnalyticEuropeanForwardEngine for option on comm future
Up until now, the exercise date for a European option on a future was
approximately equal to the underlying future expiry date. There are
cases for power derivatives where the exercise date can be a month
or so before the future expiry date so we need to use the
AnalyticEuropeanForwardEngine to generate the right future price for the
Black formula.