[Touch:691]For price strike index CDS options, price based volatilities.
If the StrikeType is Price in the trade XML, we assume the market
volatility surface is given with price along the strike dimension and
that it is a Price volatility. This is what we consume from Markit for
the conventionally price quoted indices, only CDX HY currently.

The calculation of the forward front end protection adjusted price in
this commit probably needs to be reviewed. Dividing the full notional
seems wrong.

Would need to check also that sensitivity to the price based volatility
is ok for CDX HY in the CRIF. Maybe we need to add back the conversion
from price to strike but we would then also need to update the market
data that we consume from Markit i.e. create a spread based volatility
surface.