[touch:1773]avoid the need for historical fixings in bma helper
We have to check the actual conventions of SIFMA basis swaps and possibly add required historical fixings to the
required market fixings function. This is the contents of the FIXME in the code. For now we just make it more robust so
that trades run through and do not cause a missing fixing error during curve bootstrap.