[Touch:691]Include more information on defaults.
Not all engines will use all of these fields. The main fields are prior
weight, recovery rate (i.e. final auction price) and the auction date.
The auction date is important because it is on the day after an auction
date (approx) that the new version of the index begins trading.