[Touch:691]Separate out the single name CDS option engine.
Cleaner to have the single name CDS option engine separate from the
index CDS option engine.

Restrict the strike type on single name CDS options to Spread for now.
This is the standard. Make additions based on real requirements.

Deal with non-knockout payment i.e. if knocksOut is false and the option
is a Payer.