Refactor Asian option engine builder
This changes the Asian option engine builders to reuse the existing getBlackScholesProcess(...) function instead of using a makeshift version. This enables us to remove the strike argument to the builder and engine key.

Also fixes a bug where the geometric average price MC engine was tied to the arithmetic equivalent.