Add builder for equity moneyness surface, update strike
This updates the market datum parser for EQUITY_OPTION to enable reading the new strike class used for commodity options. We still preserve the ability to read option prices but the implementation is made rather messy.

A builder is added for the BlackVarianceSurfaceMoneyness and the BlackVolatilitySurfaceMoneyness previously committed. We also add the ability to toggle between these two, reusing the same config. The implementation builds on the one for commodity vol curves with moneyness surfaces, though adapted to suit equity specific things.

The added test for the BlackVolatilitySurfaceMoneyness class ensures proper interpolation and we show that the tests in equitymarketdata.cpp can correctly parse the new strike setup, although we refrain from updating all previous tests.