[touch:1917]ensure models are calibrated before T0 valuation
The static backtest uses several runs of buildCube() and we observe that the T0 values are not always the same for the
scripted autocallable fx trade in the backtest regression. After this fix, the T0 values are the identical in each run
as expected. We still have to investigate why this seemed to work fine in 1.0.48.2. The behaviour seems to have changed
in the commit for the ql upgrade 1.18 => 1.22, see the ticket for details.