[touch:1918] Add n-th weekday for option expiry and a from future expiry independent frequency.
To calculate a option expiry relatively to a future expiry it will use the option expiry
frequency and one of the following rules:

n-th weekday: of the month of the corresponding future expiry Date (OptionNthWeekday)
day of month: in the the month of the corresponding future expiry Date (OptionExpiryDay)
BusinessDaysBefore:  n business days before the corresponding future expiry Date (OptionExpiryOffset)

If the future frequency is daily but the option frequency isn't, the anchorDate is required, to
calculate the future expiry date following the option frequency.

It's allows at most one of these nodes in the convention xml.