[touch:2058]fix - this seems to be a ql bug actually
When running the regression test CRIF_20191230_EUR/Inputs/FX_DoubleBarrierOption_Touched_Inverted I get a "barriers are
already touched" exception - this is because we trigger the calculation now where we did not before. But the calculation
should be triggered and the right place would be Instrument::additionalResults() I think. Opened a QL ticket for
clarification.