[touch:2084]all times <= 0 represent a deterministic state
We set them all to 0 to fix the consistency check now, e.g. a time -0.2 and 0.0 are consistent, but the check fails at
the moment. Since the time of a rv is meant to represent the "t of the filtration at t" a value of zero is more
consistent anyhow, since the filtration is defined only for non-negative t.