QPR-11233 make unit tests run
There is a remaining inconsistency between some pricing engines (e.g. analytic european engine) and our fx spot / rate
setup: the analytic european engine will compute the forward as the spot that is passed in (this is now the lagged spot,
before it was the spot as of today) and compound that from today to the option expiry. Whereas the compounding should be
done between today + lag and option expiry + lag, lag = fx spot lag. Nevertheless the new setup seems to be less
confusing and more correct in a sense with the remaining possible TODO to refine the pricing engines as well, where necessary.