QPR-11510 ensure monotonic times
In particular this handles the situation where "fixingDate < exerciseDate <= accrualStartDate" for an Ibor-like
coupon. We need to assume the fixing date to lie on the exerciseDate in this case to make the numerical 1f method
work. This was actually done in the previous version of the swaption pricer but got lost durint the overhaul for RFR
handling.