QPR-13487 missunderstood this code, reinstate
for global calibration where this input curve would be given, but not yet available, we would need another boolean flag
indicating whether this curve is given or not independently of whether the handle is empty - but it does not seem to be
worth adding this now, because we pass an empty handle anyhow in the curve construction. we can possibly remove these
curves from the ctor input entirely?