Add Boost libraries missing in link step.
When find_package uses config search mode for Boost instead of the old
module search mode, there are Boost libraries missing in the link step:
- Boost::log_setup missing for QLE and ORED test suites.
- Boost::regex missing in ORE-SWIG build.

To build using the CMake config file provided by Boost, you do not
create the BOOST and BOOST_LIB64 environment variables and hence
BOOST_INCLUDEDIR and BOOST_LIBRARYDIR are not set in the presets.
You instead supply the path to BoostConfig.cmake in the variable
Boost_DIR. Also, if you are using CMake 3.30 and above, you should
set the variable CMAKE_POLICY_DEFAULT_CMP0167 to NEW to avoid
warnings about the policy not being set.

This issue is not apparent when you create the BOOST and BOOST_LIB64
environment variables because CMake uses the module mode search and the
CMake FindBoost.cmake module has entries of the form:
  set(_Boost_LOG_DEPENDENCIES log_setup regex ...)
so the dependencies on log_setup and regex are pulled in.

Note: to see the difference in the link commands, you need to keep the
.rsp files that are used to store the .obj and .lib file arguments. If
building with Ninja, you can do this by passing `-d keeprsp` to the
build step. From command line, something like this for example:
  cmake --build --preset=windows-ninja-x64-debug
    --target quantext-test-suite.exe --verbose -- -d keeprsp
or if you want to add it to the preset file for VS to pick up:
  "nativeToolOptions": [ "-d keeprsp" ]