Respect the BOOST_ALL_NO_LIB variable setting.
As is done in QuantLib, if BOOST_ALL_NO_LIB is set, do not include any
auto_link.hpp files. In other words, auto linking is turned off and
the `/DEFAULTLIB:...` linker directives specifying library names to link
against are not written to the .obj files during an MSVC build.