Prevent gcc error when -Werror=return-type.
This is done a couple of other places:
  OREData/ored/portfolio/builders/yoycapfloor.cpp
  OREData/ored/portfolio/builders/capfloor.cpp

My gcc compiler version:
  gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0

If you preprocess the file first and then remove the do {...}
while(false); surrounding the code, the compilation works even with the
break; statement.

Everywhere in QuantLib that we have a non-void function with returns
inside switch case blocks, the default case has QL_FAIL not followed by
a break;. I guess it is for this reason.