Prevent ZCIIS helper observing eval date.
Adding the start date to the ZeroCouponInflationSwapHelper sets
updateDates_ to false in RelativeDateBootstrapHelper and means that the
helper does not observe Settings evaluationDate.

This brings the test in line with the way that the helper is created
in InflationCurve::buildZeroInflationCurve in OREData also.

There is no need to explicitly unregister the evaluationDate after
creation of the helper with an explicit non-default constructed start
date.

Note: YearOnYearInflationSwapHelper should probably be updated in
QuantLib to have the same behaviour i.e. that if an explicit start date
is provided, updateDates_ is set to false in RelativeDateBootstrapHelper
and the helper does not observe Settings evaluationDate. Until then, we
still need the explicit unregsiter there in
InflationCurve::buildYoYInflationCurve in OREData.