Temporary fix to allow scen gen tests to run.
This allows all scen gen tests except testCpiSwapExposure test to run.
That test still fails because the grid has only one tenor point at 5Y
and the zero inflation curve is recalculated with evaluation date set to
that date and fails because fixings are missing. The recalculation
should not be happening - it is not intended for this inflation curve
that is used in the DK inflation scenario generation to recalculate on
future dates.