Fix DK inflation simulation in orea tests.
Currently the DK inflation model is not using the t_0 zero inflation
term structure properly when calculating future scenarios.

This fix only fixes the scenario generation in the orea test suite. It
is up for discussion whether similar changes should be applied at the
point where the zero inflation term structure is created in ORE Data. I
will open a separate issue for this.

What was happening before this fix is that the zero inflation term
structure, due to a series of notifications, was being forced to
recalculate on the first simulation date for every sample and this
recalculated curve was being used in the generation of the CPI index
scenario values rather than the t_0 calculated curve.

On sample 0, there is a call to `ScenarioSimMarket::updateScenario(
const Date& d)` which changes the evaluationDate to the first grid
date. The inflation index in the ZCIIS inside the helper is observing
this evaluationDate and the inflation curve is observing the index in
turn. Due to this, the inflation curve is forced to recalculate on the
first grid date t_1 for sample 0. This t_1 recalculated curve is used
for the generation of all scenarios for t_2,...,t_N for sample 0. Note
that if relative to date t_1, some of the inflation helpers in the curve
require a CPI fixing, e.g. t_1 >= fixing date of some helper, the test
fails with exception from iterative bootstrap about missing fixing.

On samples > 0, the `applyScenario(scenario);` in
`ScenarioSimMarket::updateScenario(const Date& d)` triggers a
notification chain that leads to the inflation curve being recalculated
on t_1 for every sample. The chain is:
- ScenarioSimMarket simData_ setValue on CPI index quote
- InflationIndexObserver which in update() calls setFixing()
- adds fixings in IndexManager which in turn calls
  `notifier(name)->notifyObservers();`
- the indices in the ZCIIS swap instruments in the ZC inflation curve
  helpers are observing the notifier => they get notified
- the ZC inflation curve is observing the indices in the previous bullet
  => the curve gets notified and in its `update()` call sets
  `calculated_` to `false` => recalculation on t_1 for next sample.