Fix crashing ObservationModeTest/testDefer test.
Without the changes here, YearOnYearInflationSwapHelper observes
Settings evaluationDate. Because of this
YearOnYearInflationSwapHelper::initializeDates() gets called when
the evaluation date changes. This leads to the yyiis_ member getting
reassigned to a new YOY inflation swap. However, the original pointee is
still in deferredObervers_ of ObservableSettings. The testDefer test
then crashes when deferredObserver->update(); is called in
ObservableSettings::enableUpdates() with null deferredObserver.

Note: should open a PR in QuantLib so that when swap start is provided
      to YearOnYearInflationSwapHelper, updateDates_ gets set to false
      in RelativeDateBootstrapHelper similar to ZCIIS helper.